+++ js/src/dapps/registry/reducers.js
@@ -1,4 +1,5 @@
-import accountsReducer from './accounts/reducers.js';
+import accountsReducer from './addresses/accounts-reducer.js';
+import contactsReducer from './addresses/contacts-reducer.js';
 import lookupReducer from './Lookup/reducers.js';
 import eventsReducer from './events/reducers.js';
 import registerReducer from './register/reducers.js';
@@ -14,6 +15,7 @@ const ownerReducer = (state = null, action) =>
 
 const initialState = {
   accounts: accountsReducer(undefined, { type: '' }),
+  contacts: contactsReducer(undefined, { type: '' }),
   contract: contractReducer(undefined, { type: '' }),
   fee: feeReducer(undefined, { type: '' }),
   owner: ownerReducer(undefined, { type: '' }),
@@ -24,6 +26,7 @@ const initialState = {
 
 export default (state = initialState, action) => ({
   accounts: accountsReducer(state.accounts, action),
+  contacts: contactsReducer(state.contacts, action),
   contract: contractReducer(state.contract, action),
   fee: feeReducer(state.fee, action),
   owner: ownerReducer(state.owner, action),
